feat: To Do 삭제 버튼 클릭 시, AlertDialog를 호출하도록 구현

+++ app/toDo/data-table.tsx
@@ -15,12 +15,23 @@ import {
   TableHeader,
   TableRow,
 } from "@/components/ui/table";
-import { Button } from "@/components/ui/button";
+import { Button, buttonVariants } from "@/components/ui/button";
 import { Trash2 } from "lucide-react";
 import { ToDoType } from "@/type/toDo";
 import { useSetAtom } from "jotai";
 import { removeTodoAtom } from "@/store/toDoListAtom";
 import { Badge } from "@/components/ui/badge";
+import {
+  AlertDialog,
+  AlertDialogAction,
+  AlertDialogCancel,
+  AlertDialogContent,
+  AlertDialogDescription,
+  AlertDialogFooter,
+  AlertDialogHeader,
+  AlertDialogTitle,
+  AlertDialogTrigger,
+} from "@/components/ui/alert-dialog";
 
 interface DataTableProps<TData, TValue> {
   columns: ColumnDef<TData, TValue>[];
@@ -94,13 +105,37 @@ export function DataTable<TData, TValue>({
                   </TableCell>
 
                   <TableCell className="w-[36px]">
-                    <Button
-                      variant="ghost"
-                      size="icon"
-                      onClick={() => handleDeleteButtonClick(rowData.id)}
-                    >
-                      <Trash2 />
-                    </Button>
+                    <AlertDialog>
+                      <AlertDialogTrigger>
+                        <Button
+                          variant="ghost"
+                          size="icon"
+                          className="w-[24px] h-[24px] p-[4px]"
+                          asChild
+                        >
+                          <Trash2 />
+                        </Button>
+                      </AlertDialogTrigger>
+                      <AlertDialogContent>
+                        <AlertDialogHeader>
+                          <AlertDialogTitle>Delete</AlertDialogTitle>
+                          <AlertDialogDescription>
+                            Are you sure you want to delete this to do?
+                          </AlertDialogDescription>
+                        </AlertDialogHeader>
+                        <AlertDialogFooter>
+                          <AlertDialogCancel>Cancel</AlertDialogCancel>
+                          <AlertDialogAction
+                            className={buttonVariants({
+                              variant: "destructive",
+                            })}
+                            onClick={() => handleDeleteButtonClick(rowData.id)}
+                          >
+                            Delete
+                          </AlertDialogAction>
+                        </AlertDialogFooter>
+                      </AlertDialogContent>
+                    </AlertDialog>
                   </TableCell>
                 </TableRow>
               );
